pyproject.toml
[project]
name = "opengrader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "pypdf>=6.7.5",
    "ruamel-yaml>=0.19.1",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
